day-1-foundations/requirements.txt
openai
groq
google-generativeai
mistralai

langchain
langchain-community

tiktoken
pydantic

faiss-cpu
chromadb
sentence-transformers

python-dotenv
requests
tqdm
langchain_huggingface